requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
distro==1.9.0
fastapi==0.136.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
magic-filter==1.0.12
multidict==6.7.1
openai==2.32.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
sniffio==1.3.1
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
yarl==1.23.0
